#8 df: Validation seems to work now
