+++ src/components/scenario_explorer/ScenarioExplorerTree.jsx
@@ -3,7 +3,7 @@ import ExpandableScenarioNode from './ScenarioExplorerNode.jsx';
 import EditScenarioTestMenu from '../EditScenarioTestMenu';
 import { useState } from 'react';
 import axios from 'axios';
-import ScenarioTestMenu from "../ScenarioTestMenu.jsx";
+import ScenarioTestMenu from "../scenariotest/ScenarioTestMenu.jsx";
 import ReactDOM from "react-dom/client";
 import {RouterProvider} from "react-router-dom";
 export default function ScenarioTree(props) {
